feat(finder): drag files from explorer to open in editor (T630 P2)
Hoist a single DndContext up to FinderShell, spanning both the explorer and
the editor. Explorer file rows become useDraggable sources (data.type
explorer-file + node); TabHost is the useDroppable editor zone, lit with a
ring on hover. Shell onDragEnd dispatches: an explorer-file dropped over the
zone or any tab opens it pinned; otherwise a tab drag reorders. TabStrip drops
its own DndContext, keeping only the SortableContext under the hoisted one.
Folders stay non-draggable; 4px activation distance keeps click/open intact.
